Refactor mailer blacklist from trait to DI [MAILPOET-2208]
